requirements.txt
Django==5.0.6
psycopg2==2.9.9
requests==2.32.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
pytest ==8.2.0
pytest-django==4.8.0

